[api][webui] Move Project query to it's own class
